+++ contracts/interfaces/IOrionConfig.sol
@@ -100,13 +100,13 @@ interface IOrionConfig {
 
     /// @notice Returns the total number of whitelisted assets
     /// @return The count of whitelisted assets
-    function whitelistedAssetsLength() external view returns (uint256);
+    function whitelistedAssetsLength() external view returns (uint16);
 
     /// @notice Returns the whitelisted asset address at the specified index
     /// @dev Uses EnumerableSet ordering, which may change when assets are added/removed
     /// @param index The index of the asset to retrieve
     /// @return The address of the whitelisted asset at the given index
-    function getWhitelistedAssetAt(uint256 index) external view returns (address);
+    function getWhitelistedAssetAt(uint16 index) external view returns (address);
 
     /// @notice Returns all whitelisted assets
     /// @return An array of whitelisted asset addresses
