+++ contracts/interfaces/IOrionTransparentVault.sol
@@ -27,7 +27,7 @@ interface IOrionTransparentVault is IOrionVault {
     /// @notice Get the transparent intent.
     /// @return tokens The tokens in the intent.
     /// @return weights The weights in the intent.
-    function getIntent() external view returns (address[] memory tokens, uint256[] memory weights);
+    function getIntent() external view returns (address[] memory tokens, uint32[] memory weights);
 
     /// @notice Updates the vault's portfolio state and total assets
     /// @dev Can only be called by the liquidity orchestrator.
